BDD/bdd_project/requirements.txt
allure-python-commons
selenium
behave
webdriver-manager
allure-behave
pytest
openpyxl
configparse
gherkin-official